pyproject.toml
[tool.poetry]
name = "npg_notify"
version = "0.0.1"
description = "Utility for client notifications"
authors = ["Marina Gourtovaia <mg8@sanger.ac.uk>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = { version = "^2.0.1", extras = ["pymysql"] }
SQLAlchemy-Utils = "^0.41.2"
cryptography = { version = "^41.0.3" }

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
PyYAML = "^6.0.0" 
ruff = "^0.4.9"

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  "src"
]
